# apps/backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Personal finance tracking: transactions, categories, dashboard and cash flow reports.",
        "title": "Financial Management API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/telegram/link": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Telegram"
                ],
                "summary": "Telegram link status",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_rezanii_tracking-cashflow_apps_backend_internal_utils.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_rezanii_tracking-cashflow_apps_backend_internal_dto.TelegramLinkResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_rezanii_tracking-cashflow_apps_backend_internal_utils.Envelope"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Telegram"
                ],
                "summary": "Unlink Telegram chat",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_rezanii_tracking-cashflow_apps_backend_internal_utils.Envelope"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_rezanii_tracking-cashflow_apps_backend_internal_utils.Envelope"
                        }
                    }
                }
            }
        },
        "/telegram/pairing-code": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns a short-lived single-use code. Send \"/start \u003ccode\u003e\" to the bot from the\nchat that should receive reports. Until a chat is paired the bot answers nothing.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Telegram"
                ],
                "summary": "Issue Telegram pairing code",
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_rezanii_tracking-cashflow_apps_backend_internal_utils.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_rezanii_tracking-cashflow_apps_backend_internal_dto.TelegramPairingCodeResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_rezanii_tracking-cashflow_apps_backend_internal_utils.Envelope"
                        }
                    }
                }
            }
        },
        "/telegram/send/daily-report": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Telegram"
                ],
                "summary": "Send daily report to Telegram",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Report date, YYYY-MM-DD. Defaults to today.",
                        "name": "date",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_rezanii_tracking-cashflow_apps_backend_internal_utils.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_rezanii_tracking-cashflow_apps_backend_internal_dto.TelegramSendResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "422": {
                        "description": "No chat is linked",
                        "schema": {
                            "$ref": "#/definitions/github_com_rezanii_tracking-cashflow_apps_backend_internal_utils.Envelope"
                        }
                    }
                }
            }
        },
        "/telegram/webhook": {
            "post": {
                "description": "Called by Telegram, not by clients. The X-Telegram-Bot-Api-Secret-Token header\nmust match TELEGRAM_WEBHOOK_SECRET, so knowing the URL is not enough to post\nforged updates.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Telegram"
                ],
                "summary": "Telegram webhook",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_rezanii_tracking-cashflow_apps_backend_internal_utils.Envelope"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_rezanii_tracking-cashflow_apps_backend_internal_utils.Envelope"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_rezanii_tracking-cashflow_apps_backend_internal_dto.TelegramLinkResponse": {
            "type": "object",
            "properties": {
                "chat_id": {
                    "type": "integer",
                    "example": 123456789
                },
                "chat_title": {
                    "type": "string",
                    "example": "Reza"
                },
                "linked": {
                    "type": "boolean",
                    "example": true
                },
                "linked_at": {
                    "type": "string"
                },
                "username": {
                    "type": "string",
                    "example": "rezani"
                }
            }
        },
        "github_com_rezanii_tracking-cashflow_apps_backend_internal_dto.TelegramPairingCodeResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "7F3K9Q"
                },
                "expires_at": {
                    "type": "string"
                },
                "instruction": {
                    "description": "Instruction is ready to show in the UI so the user does not have to be told the\ncommand format separately.",
                    "type": "string",
                    "example": "Send /start 7F3K9Q to the bot"
                }
            }
        },
        "github_com_rezanii_tracking-cashflow_apps_backend_internal_dto.TelegramSendResponse": {
            "type": "object",
            "properties": {
                "characters": {
                    "description": "Characters is the rendered length before splitting, useful when a report grows near\nthe 4096-character limit.",
                    "type": "integer",
                    "example": 2480
                },
                "chat_id": {
                    "type": "integer",
                    "example": 123456789
                },
                "messages": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
...
        "github_com_rezanii_tracking-cashflow_apps_backend_internal_utils.Envelope": {
            "type": "object",
            "properties": {
                "data": {},
                "errors": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Send the access token as: Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}